Move test fixtures to a common location within the modules package
